refactor(Navigation) : Change selected product price data type.

+++ app/src/main/java/com/example/itgproductapiapp/ui/navigation/MainNavDestinations.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class MainScreen(
     val name: String? = null,
-    val price: Double? = null,
+    val price: Int? = null,
     val imageUrl: List<String>? = null,
     val category: String? = null
 )
